requirements_base.txt
otree-core[mturk]>=1.4.28
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
